requirements.txt
requests==2.31.0
urllib3==2.2.1
Flask==3.0.3
pillow==10.3.0
transformers==4.40.2
torchvision==0.17.0
torch==2.2.0
nltk==3.8.1
easyocr==1.7.1
pymystem3==0.2.0
catboost==1.2.5
numpy==1.26.4